refactor(bench): split runner helpers
